Shop/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
boto3==1.26.37
botocore==1.29.37
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.4
django-admin-interface==0.24.2
django-celery-beat==2.4.0
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-nose==1.4.7
django-rest-framework==0.1.0
django-storages==1.13.2
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kavenegar==1.1.2
kombu==5.2.4
MarkupSafe==2.1.1
nose==1.3.7
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
prompt-toolkit==3.0.36
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
